classifier/requirements.txt
colorama
google-cloud-functions
google-cloud-logging
h5py
natsort
pgb-broker-utils==0.2.43
scikit-learn
seaborn
supernnova==1.3
torch
# https://download.pytorch.org/whl/cpu/torch-1.9.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
# cloud fncs don't support GPUs
# https://stackoverflow.com/questions/55449313/google-cloud-function-python-3-7-requirements-txt-makes-deploy-fail
tqdm
